# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mini ERP API
  version: "1.0.0"
servers:
  - url: http://localhost:5000/api/v1
paths:
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200': { description: JWT token }
  /auth/register:
    post:
      summary: Register new user (Staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '200': { description: User registered }

  /products:
    get:
      summary: List products
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: OK }
    post:
      summary: Create product
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCreateDto' }
      responses:
        '201': { description: Created }
  /products/{id}:
    get:
      summary: Get product by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }

  /inventory/{productId}:
    get:
      summary: Get inventory by product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
  /inventory/adjust:
    post:
      summary: Adjust stock
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: productId
          schema: { type: integer }
        - in: query
          name: warehouseId
          schema: { type: integer }
        - in: query
          name: quantity
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string, enum: [IN, OUT, ADJUST] }
        - in: query
          name: reference
          schema: { type: string }
      responses:
        '200': { description: OK }

  /salesorders:
    post:
      summary: Create sales order
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SalesOrderCreateDto' }
      responses:
        '201': { description: Created }
  /salesorders/{id}:
    get:
      summary: Get sales order by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
  /salesorders/{id}/ship:
    post:
      summary: Ship sales order (reduce stock)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }

  /purchaseorders:
    post:
      summary: Create purchase order
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PurchaseOrderCreateDto' }
      responses:
        '201': { description: Created }
  /purchaseorders/{id}:
    get:
      summary: Get purchase order by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
  /purchaseorders/{id}/receive:
    post:
      summary: Receive purchase order (increase stock)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginDto:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
    RegisterDto:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
    ProductCreateDto:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        categoryId: { type: integer }
        imageUrl: { type: string }
    SalesOrderCreateDto:
      type: object
      properties:
        customerId: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesOrderItemDto'
    SalesOrderItemDto:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        unitPrice: { type: number }
    PurchaseOrderCreateDto:
      type: object
      properties:
        supplierId: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseOrderItemDto'
    PurchaseOrderItemDto:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        unitPrice: { type: number }